AuthDatabase have finally arrived
Changes in ClientSessionStore
  It uses default std::mutex and std::lock_guard instead of ACE stuff
  It now handles the reaping of stale connections almost on it's own  ( still need an EventProcessor to register the timer to itself, and call `reap_stale_links`
  Rename all ClientSessionStore methods to snake_case
  Created helper methods `mark_session_for_reaping` `unmark_session_for_reaping`
  To allow session object reuse for re-connected players it also provides `create_or_reuse_session_for`
  Introduce ReaperProtection template class that uses RAII idiom to protect given session from reaping for the life time of ReaperProtection
  Temporary sessions marked for reaping will be removed even if the `link` variable is still set in the session, an evFinish will be posted the link

Changes to InternalEvents
  ReloadConfig event was added, every server should handle this event to perform stop/load config/start set of operations
  ServiceStatus message, used by servers to broadcast changes in their state on MessageBus was added

Removed classes/files
  ServerManager - all of it's functionality has been assumed by HandlerLocator and the main.cpp
  AdminServerInterface - no longer used anywhere in the codebase
  AdminDatabase - functionality taken over by AuthDatabase service.

Added AuthDatabase and GameDatabase services
  For now only AuthDatabase is used, the GameDatabase will replace CharacterDB usages in the next PR.

Changes related to MapServer
  Add initial map server/instance registry that takes account of map names

Changes in GameServer
  Many cosmetic changes
  Move GameLinkEndpoint definition from header to anonymous namespace in GameServer cpp file
  GameHandler is created and activated in the GameServer constructor now

Changes in AuthServer/Handler
  AuthHandler now asks AuthDatabase about the account details, and waits for its response
    Timeouts are handled by marking the un-authorized sessions as temporary, if database times-out,
    the session is simply reaped and the link gets evFinish message, to force it to close up.
  AuthServer is now owner of the AuthHandler, and manages it's whole lifetime.

Main.cpp
  Creates the MessageBus instance, and registers it with HandlerLocator
  Adds TIMED_LOG macro to display server boot-up times
  Creates and starts all services.
  Introduces MessageBusMonitor which handles service status messages, and shuts down when any service fails.
Settings.cpp/.h
  Since we now read settings from different threads, and QSettings is reentrant, but not multi-threaded, we now create QSettings based on Settings provided path

Misc stuff:

* Move `info` method from SEGSEvent header to cpp file
+ Add a `useless-cast` warning and `pedantic` flags on gcc-like compilers
- Remove forced `std=c++11` from gcc commandline, CMake handles this with CMAKE_CXX_STANDARD
* AuthLink instances now have their own 'local' copy of their target handler. This helps with proper construction, non-server use, maybe handling multiple AuthServer connections ?
* AuthExerciser makes use of AuthConnector which overrides make_svc_handler to properly construct the AuthLink objects
* All servers now are EventProcessors ( mainly used to properly shutdown/reload configurations )

+++ Projects/CoX/Common/NetStructures/Character.cpp
@@ -117,7 +117,7 @@ void Character::SendCharBuildInfo(BitStream &bs) const
         {
             uint32_t num_powers=m_powers.size();
             bs.StorePackedBits(5,0);
-            bs.StorePackedBits(4,uint32_t(num_powers));
+            bs.StorePackedBits(4,num_powers);
             for(const CharacterPower &power : m_powers)
             {
                 //sendPower(bs,0,0,0);
